test: add inner class with configuration
